Change logging on all containers to "local."
Log file size is unbounded by default, and we log a lot of stuff.
The "local" setting for logs uses a max file size for the log file,
which will limit how much disk space is used by our logs.

All containers are set to use a 10MB log, which should give us
plenty of historical data to see what is happening.